feat(event-submission): new event modal component
I moved the stuff from the /event-list.svelte component (the button, mainly) into a new event-submission modal so that they would be co-located with the actual modal. I'm not sure if this is the right design decision, but hey, it works.

I also changed it to say "Suggest an Event" instead of "Add an Event" because I thought it was a better word choice.

Also I'm not even sure that event-submission is even a good name for that component.