refactor(contact): add emails and remove form
This should hopefully be more clear for users as having a form that just
sends an email is kinda pointless.

The only real benifit of having the form is that we can just have a
discord webhook, but that's probably not what we should do in the long
term anyways.